Add Text DB and Interaction with BL (+DB Testing)

+++ src/Database/textDB.java
@@ -4,35 +4,42 @@ import com.company.Cell;
 import com.company.DBInterfaceIn;
 import com.company.DBInterfaceOut;
 import com.company.GameOfLife;
-
-import java.awt.*;
 import java.io.*;
 import java.util.Enumeration;
 import java.util.Hashtable;
-import java.util.Scanner;
 
-public class textDB implements DBInterfaceOut
+
+public class TextDB implements DBInterfaceOut
 {
-    DBInterfaceIn gameControls;
-    public textDB(GameOfLife g)
+
+    protected DBInterfaceIn gameControls;
+    public TextDB(GameOfLife g)
     {
         gameControls = g;
         g.addDBListener(this);
     }
     public void deleteRecentState()                        //Delete The Recent State
     {
-        FileWriter fwOb = null;
         BufferedReader br;
         File obj;
-        String data,LastFile=null;
-        try {
+        String data, recentSavedFile=null;
+        File f = new File("SaveStates.txt");
+        if(!f.exists() )
+        {
+            return;
+        }
+        try
+        {
             br = new BufferedReader(new FileReader("SaveStates.txt"));
             while ((data = br.readLine()) != null)
             {
-                LastFile = data;
+                recentSavedFile = data;
             }
-            DeleteLastLine();
-            obj = new File(LastFile);
+            br.close();
+            if(recentSavedFile==null)
+                return;
+            DeleteLine(recentSavedFile);
+            obj = new File(recentSavedFile);
             obj.delete();
         }
         catch (IOException e)
@@ -40,36 +47,8 @@ public class textDB implements DBInterfaceOut
             e.printStackTrace();
         }
     }
-    public static void DeleteLastLine()
-    {
-        RandomAccessFile f = null;
-        try
-        {
-            f = new RandomAccessFile("SaveStates.txt", "rw");
-            long length = f.length() - 1;
-            byte b;
-            do
-            {
-                length -= 1;
-                f.seek(length);
-                b = f.readByte();
-            }
-            while (b != 10);
-            f.setLength(length + 1);
-            f.close();
-        }
-        catch (FileNotFoundException e)
-        {
-            e.printStackTrace();
-        }
-        catch (IOException e)
-        {
-            e.printStackTrace();
-        }
-    }
 
-
-    public static  void  DeleteBetweenLine(String name)
+    public static  void  DeleteLine(String name)
     {
         File inputFile = new File("SaveStates.txt");
         File tempFile = new File("myTempFile.txt");
@@ -78,20 +57,20 @@ public class textDB implements DBInterfaceOut
             BufferedReader reader = new BufferedReader(new FileReader(inputFile));
             BufferedWriter writer = new BufferedWriter(new FileWriter(tempFile));
 
-            String lineToRemove = name;
             String currentLine;
 
-            while((currentLine = reader.readLine()) != null)
+            while ((currentLine = reader.readLine()) != null)
             {
                 // trim newline when comparing with lineToRemove
                 String trimmedLine = currentLine.trim();
-                if(trimmedLine.equals(lineToRemove)) continue;
+                if (trimmedLine.equals(name))
+                    continue;
                 writer.write(currentLine + System.getProperty("line.separator"));
             }
             writer.close();
             reader.close();
             inputFile.delete();
-            boolean successful = tempFile.renameTo(inputFile);
+            tempFile.renameTo(inputFile);
         }
         catch (FileNotFoundException e)
         {
@@ -101,24 +80,19 @@ public class textDB implements DBInterfaceOut
         {
             e.printStackTrace();
         }
-    }
 
-    //delete specific state
-    public void deleteSelectedState(String name)                        //Delete The Given State
+    }
+    public void deleteState(String name)                        //Delete The Given State
     {
-        DeleteBetweenLine(name);
+        DeleteLine(name);
         File obj = new File(name);
         obj.delete();
     }
-
-    //save State into text file
-    public void saveState(Hashtable ht,String Name)                 //SAVE STATE IN FILE
+    public void saveState(Hashtable ht,String name)                 //SAVE STATE IN FILE
     {
-        Cell c;
+        Cell cell;
         File file_obj = new File("SaveStates.txt");
         BufferedWriter out;
-        BufferedWriter out1;
-        Desktop desktop = Desktop.getDesktop();
         Enumeration e = ht.elements();
         int x_axis,y_axis;
         try
@@ -128,7 +102,7 @@ public class textDB implements DBInterfaceOut
                 try
                 {
                     out = new BufferedWriter(new FileWriter("SaveStates.txt",true));
-                    out.write(Name+".txt");
+                    out.write(name+".txt");
                     out.append('\n');
                     out.close();
                 }
@@ -143,7 +117,7 @@ public class textDB implements DBInterfaceOut
                 try
                 {
                     out = new BufferedWriter(new FileWriter("SaveStates.txt"));
-                    out.write(Name+".txt");
+                    out.write(name+".txt");
                     out.append('\n');
                     out.close();
                 }
@@ -159,23 +133,13 @@ public class textDB implements DBInterfaceOut
         }
         try
         {
-            file_obj = new File(Name+".txt");
-
-        }
-        catch (Exception exception)
-        {
-            exception.printStackTrace();
-        }
-        try
-        {
-            out = new BufferedWriter(new FileWriter(Name+".txt"));
+            out = new BufferedWriter(new FileWriter(name+".txt"));
             while(e.hasMoreElements())                                     //WRITING TO TEXT FILE ALL HASHTABLE
             {
+                cell = (Cell) e.nextElement();
 
-                c = (Cell) e.nextElement();
-
-                x_axis = c.x_axis;
-                y_axis = c.y_axis;
+                x_axis = cell.x_axis;
+                y_axis = cell.y_axis;
 
                 out.write(String.valueOf(x_axis));
                 out.append(',');
@@ -190,65 +154,71 @@ public class textDB implements DBInterfaceOut
             exp2.printStackTrace();
         }
     }
-
-    //load last Added State
-    public  Hashtable loadRecentState()
+    public  Hashtable<Cell, Cell> loadRecentState()
     {
         BufferedReader br;
-        Hashtable<Cell,Cell> ht =new Hashtable<Cell, Cell>();
-        File obj;
-        String data,LastFile=null;
-        int X_Axis = 0,Y_Axis=0,val=0,size;
-        int array1[] = new int[200];
+        Hashtable<Cell, Cell> ht =new Hashtable<>();
+        String data,recentSavedFile=null;
+        int x_Axis = 0, y_Axis=0, val=0;
+        int[] array1 = new int[200];
+        File f = new File("SaveStates.txt");
+        if(!f.exists() )
+        {
+            return ht;
+        }
         try
         {
             br = new BufferedReader(new FileReader("SaveStates.txt"));     //GETTING RECENT NAME OF FILE
             while ((data = br.readLine()) != null)
             {
-                LastFile = data;
+                recentSavedFile = data;
             }
+            br.close();
         }
         catch (IOException e)
         {
             e.printStackTrace();
         }
+        if(recentSavedFile==null)
+            return ht;
         try
         {
-            br = new BufferedReader(new FileReader(String.valueOf(LastFile)));
+            br = new BufferedReader(new FileReader(recentSavedFile));
             while((data = br.readLine()) != null)
             {
-                for( int i = 0; i<=data.length();i++)                              //EXTRATCING FROM RECENT FILE
+                for( int i = 0; i<=data.length();i++)                              //Extracting FROM RECENT FILE
                 {
                     if(i==data.length())
                     {
                         for (int j = 0; j < val; j++) {                            //CONVERTING INT ARRAY TO INTEGER
-                            Y_Axis = 10 * Y_Axis + array1[j];
+                            y_Axis = 10 * y_Axis + array1[j];
                         }
                         break;
                     }
                     if(Character.isDigit(data.charAt(i)))                          //STORING IN INT ARRAY
                     {
                         array1[val] = data.charAt(i)-48;
-                        //System.out.println(array1[val]);
                         val = val+1;
                     }
                     else if(data.charAt(i) == ',' )                                  //CONVERTING INT ARRAY TO INTEGER
                     {
                         for (int j = 0; j < val; j++) {
-                            X_Axis = 10 * X_Axis + array1[j];
+                            x_Axis = 10 * x_Axis + array1[j];
                         }
                         val = 0;
                     }
                 }
-                Cell c = new Cell(X_Axis, Y_Axis);                                     //STORING IN HASHTABLE
+                Cell c = new Cell(x_Axis, y_Axis);                                       //Storing IN HASHTABLE
                 ht.put(c,c);
                 val = 0;
-                X_Axis = 0;
-                Y_Axis = 0;
+                x_Axis = 0;
+                y_Axis = 0;
             }
+            br.close();
         }
         catch (FileNotFoundException e)
         {
+
             e.printStackTrace();
         }
         catch (IOException e)
@@ -257,21 +227,25 @@ public class textDB implements DBInterfaceOut
         }
         return ht;
     }
-
-    //get Names of Saved States
     public String[] getStates()
     {
-        String stateNames[] = new String[60];
+        String[] statesName = new String[60];
         int val = 0;
         BufferedReader br;
+        File f = new File("SaveStates.txt");
+        if(!f.exists() )
+        {
+            return statesName;
+        }
         try
         {
             br =new BufferedReader(new FileReader("SaveStates.txt"));
-            while((stateNames[val] = br.readLine())!=null)
+            while((statesName[val] = br.readLine())!=null)
             {
                 val = val+1;
             }
-            stateNames[val] = "/0";
+            statesName[val] = "\0";
+            br.close();
         }
         catch (FileNotFoundException e)
         {
@@ -281,21 +255,19 @@ public class textDB implements DBInterfaceOut
         {
             e.printStackTrace();
         }
-        return stateNames;
+        return statesName;
     }
 
-    //load Specific State
-    public Hashtable loadState(String name)
+    public Hashtable<Cell, Cell> loadState(String Name)
     {
         BufferedReader br;
-        Hashtable<Cell,Cell> ht =new Hashtable<Cell,Cell>();
-        File obj;
+        Hashtable<Cell,Cell> ht =new Hashtable<>();
         String data;
-        int X_Axis = 0,Y_Axis=0,val=0,size;
-        int array1[] = new int[200];
+        int X_Axis = 0,Y_Axis=0,val=0;
+        int[] array1 = new int[200];
         try
         {
-            br = new BufferedReader(new FileReader(String.valueOf(name)));
+            br = new BufferedReader(new FileReader(String.valueOf(Name)));
             while((data = br.readLine()) != null)
             {
                 for( int i = 0; i<=data.length();i++)
@@ -327,6 +299,7 @@ public class textDB implements DBInterfaceOut
                 X_Axis = 0;
                 Y_Axis = 0;
             }
+            br.close();
         }
         catch (FileNotFoundException e)
         {
